requirements.txt
discord.py==2.3.2
flask==2.3.3
werkzeug==2.3.7
gunicorn==21.2.0
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0